requirements.txt
# ── Core pipeline ─────────────────────────────────────────────────────────────
yt-dlp==2026.3.17
ffmpeg-python==0.2.0
moviepy==2.2.1
imageio-ffmpeg==0.6.0
ImageIO==2.37.0
pillow==11.3.0
numpy==2.2.6

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
tqdm==4.67.3
proglog==0.1.12
decorator==5.2.1
future==1.0.0
pytube==15.0.0

# ── Telegram bot ──────────────────────────────────────────────────────────────
python-telegram-bot==22.7
APScheduler==3.11.2
httpx==0.28.1
httpcore==1.0.9
anyio==4.13.0
h11==0.16.0
certifi==2026.2.25
idna==3.11
typing_extensions==4.15.0
exceptiongroup==1.3.1
tzlocal==5.3.1

